chore#521: 계층 책임 분리
